@@ -3,6 +3,7 @@
 	"version": "1.0.0",
 	"description": "Track calories and nutrients in the food you eat.",
 	"license": "GPL-3.0-only",
+	"type": "module",
 	"author": {
 		"name": "Jenny Belanger",
 		"email": "mail@jennybelanger.com",
@@ -13,53 +14,51 @@
 		"url": "https://github.com/jlbelanger/food.git"
 	},
 	"scripts": {
-		"start": "npx update-browserslist-db@latest && react-scripts start",
-		"build": "react-scripts build",
+		"start": "node_modules/.bin/update-browserslist-db && vite",
+		"build": "vite build",
 		"lint": "yarn lint:css && yarn lint:js",
-		"lint:css": "node_modules/.bin/stylelint 'src/scss/**/*.scss'",
-		"lint:js": "node_modules/.bin/eslint . --ext .js",
+		"lint:css": "node_modules/.bin/stylelint 'src/css/**/*.css'",
+		"lint:js": "node_modules/.bin/eslint .",
 		"test:cypress": "node_modules/.bin/cypress open"
 	},
 	"dependencies": {
-		"@jlbelanger/formosa": "jlbelanger/formosa#1061f96c17116e84f4f781e84e72d712d8623b26",
-		"chart.js": "^4.4.7",
+		"@csstools/postcss-global-data": "^3.1.0",
+		"@jlbelanger/formosa": "jlbelanger/formosa#99e1ad71849b821dea3957851b50a16ba3cbc798",
+		"@vitejs/plugin-react": "^5.1.2",
+		"chart.js": "^4.5.1",
 		"chartjs-adapter-luxon": "^1.3.1",
 		"chartjs-plugin-zoom": "^2.2.0",
-		"dotenv": "^16.4.7",
-		"get-value": "^3.0.1",
+		"get-value": "^4.0.1",
 		"js-cookie": "^3.0.5",
-		"luxon": "^3.5.0",
+		"luxon": "^3.7.2",
+		"normalize.css": "^8.0.1",
+		"postcss": "^8.5.6",
+		"postcss-preset-env": "^10.6.0",
 		"prop-types": "^15.8.1",
-		"react": "^19.0.0",
-		"react-chartjs-2": "^5.2.0",
-		"react-dom": "^19.0.0",
+		"react": "^19.2.3",
+		"react-chartjs-2": "^5.3.1",
+		"react-dom": "^19.2.3",
 		"react-promise-tracker": "^2.1.1",
-		"react-router-dom": "~5.3.4",
-		"react-scripts": "^5.0.1",
-		"sass": "^1.83.0",
-		"uuid": "^11.0.3"
+		"react-router": "^7.11.0",
+		"uuid": "^13.0.0",
+		"vite": "^7.3.0",
+		"vite-plugin-svgr": "^4.5.0"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.26.0",
-		"@babel/eslint-parser": "^7.25.9",
-		"@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-		"@babel/plugin-transform-private-property-in-object": "^7.25.9",
-		"@babel/preset-env": "^7.26.0",
-		"@babel/preset-react": "^7.26.3",
-		"@stylistic/stylelint-plugin": "^3.1.1",
-		"cypress": "^13.17.0",
+		"@eslint/js": "^9.39.2",
+		"@stylistic/eslint-plugin": "^5.6.1",
+		"@stylistic/stylelint-plugin": "^4.0.0",
+		"cypress": "^15.8.1",
 		"cypress-file-upload": "^5.0.8",
-		"eslint": "^8.57.1",
-		"eslint-config-airbnb": "^19.0.4",
-		"eslint-plugin-import": "^2.31.0",
+		"eslint": "^9.39.2",
+		"eslint-plugin-import": "^2.32.0",
 		"eslint-plugin-jsx-a11y": "^6.10.2",
-		"eslint-plugin-react": "^7.37.3",
-		"eslint-plugin-react-hooks": "^5.1.0",
-		"normalize.css": "^8.0.1",
-		"stylelint": "^16.12.0",
-		"stylelint-config-standard-scss": "^14.0.0",
-		"stylelint-order": "^6.0.4",
-		"stylelint-scss": "^6.10.0"
+		"eslint-plugin-react": "^7.37.5",
+		"eslint-plugin-react-hooks": "^7.0.1",
+		"globals": "^16.5.0",
+		"stylelint": "^16.26.1",
+		"stylelint-config-standard": "^39.0.1",
+		"stylelint-order": "^7.0.0"
 	},
 	"browserslist": [
 		"defaults",
