@@ -65,7 +65,7 @@ return [
 	|
 	*/
 
-	'timezone' => env('APP_TIMEZONE', 'UTC'),
+	'timezone' => 'UTC',
 
 	/*
 	|--------------------------------------------------------------------------
@@ -101,7 +101,7 @@ return [
 
 	'previous_keys' => [
 		...array_filter(
-			explode(',', env('APP_PREVIOUS_KEYS', ''))
+			explode(',', (string) env('APP_PREVIOUS_KEYS', ''))
 		),
 	],
 
