@@ -109,7 +109,7 @@ return [
 	| Password Confirmation Timeout
 	|--------------------------------------------------------------------------
 	|
-	| Here you may define the amount of seconds before a password confirmation
+	| Here you may define the number of seconds before a password confirmation
 	| window expires and users are asked to re-enter their password via the
 	| confirmation screen. By default, the timeout lasts for three hours.
 	|
