@@ -15,7 +15,7 @@ return [
 	|
 	*/
 
-	'default' => env('CACHE_STORE', 'file'),
+	'default' => env('CACHE_STORE', 'database'),
 
 	/*
 	|--------------------------------------------------------------------------
@@ -27,7 +27,8 @@ return [
 	| same cache driver to group types of items stored in your caches.
 	|
 	| Supported drivers: "array", "database", "file", "memcached",
-	|                    "redis", "dynamodb", "octane", "null"
+	|                    "redis", "dynamodb", "octane",
+	|                    "failover", "null"
 	|
 	*/
 
@@ -88,6 +89,14 @@ return [
 			'driver' => 'octane',
 		],
 
+		'failover' => [
+			'driver' => 'failover',
+			'stores' => [
+				'database',
+				'array',
+			],
+		],
+
 	],
 
 	/*
@@ -101,7 +110,7 @@ return [
 	|
 	*/
 
-	'prefix' => env('CACHE_PREFIX', Str::slug(env('APP_NAME', 'laravel'), '_') . '_cache_'),
+	'prefix' => env('CACHE_PREFIX', Str::slug((string) env('APP_NAME', 'laravel')) . '-cache-'),
 
 	'enable' => env('ENABLE_CACHE', true),
 
