@@ -40,6 +40,7 @@ return [
 			'busy_timeout' => null,
 			'journal_mode' => null,
 			'synchronous' => null,
+			'transaction_mode' => 'DEFERRED',
 		],
 
 		'mysql' => [
@@ -58,7 +59,7 @@ return [
 			'strict' => true,
 			'engine' => null,
 			'options' => extension_loaded('pdo_mysql') ? array_filter([
-				PDO::MYSQL_ATTR_SSL_CA => env('MYSQL_ATTR_SSL_CA'),
+				(PHP_VERSION_ID >= 80500 ? \Pdo\Mysql::ATTR_SSL_CA : \PDO::MYSQL_ATTR_SSL_CA) => env('MYSQL_ATTR_SSL_CA'),
 			]) : [],
 		],
 
@@ -78,7 +79,7 @@ return [
 			'strict' => true,
 			'engine' => null,
 			'options' => extension_loaded('pdo_mysql') ? array_filter([
-				PDO::MYSQL_ATTR_SSL_CA => env('MYSQL_ATTR_SSL_CA'),
+				(PHP_VERSION_ID >= 80500 ? \Pdo\Mysql::ATTR_SSL_CA : \PDO::MYSQL_ATTR_SSL_CA) => env('MYSQL_ATTR_SSL_CA'),
 			]) : [],
 		],
 
@@ -145,7 +146,8 @@ return [
 
 		'options' => [
 			'cluster' => env('REDIS_CLUSTER', 'redis'),
-			'prefix' => env('REDIS_PREFIX', Str::slug(env('APP_NAME', 'laravel'), '_') . '_database_'),
+			'prefix' => env('REDIS_PREFIX', Str::slug((string) env('APP_NAME', 'laravel')) . '-database-'),
+			'persistent' => env('REDIS_PERSISTENT', false),
 		],
 
 		'default' => [
@@ -155,6 +157,10 @@ return [
 			'password' => env('REDIS_PASSWORD'),
 			'port' => env('REDIS_PORT', '6379'),
 			'database' => env('REDIS_DB', '0'),
+			'max_retries' => env('REDIS_MAX_RETRIES', 3),
+			'backoff_algorithm' => env('REDIS_BACKOFF_ALGORITHM', 'decorrelated_jitter'),
+			'backoff_base' => env('REDIS_BACKOFF_BASE', 100),
+			'backoff_cap' => env('REDIS_BACKOFF_CAP', 1000),
 		],
 
 		'cache' => [
@@ -164,6 +170,10 @@ return [
 			'password' => env('REDIS_PASSWORD'),
 			'port' => env('REDIS_PORT', '6379'),
 			'database' => env('REDIS_CACHE_DB', '1'),
+			'max_retries' => env('REDIS_MAX_RETRIES', 3),
+			'backoff_algorithm' => env('REDIS_BACKOFF_ALGORITHM', 'decorrelated_jitter'),
+			'backoff_base' => env('REDIS_BACKOFF_BASE', 100),
+			'backoff_cap' => env('REDIS_BACKOFF_CAP', 1000),
 		],
 
 	],
