@@ -54,7 +54,7 @@ return [
 
 		'stack' => [
 			'driver' => 'stack',
-			'channels' => explode(',', env('LOG_STACK', 'single')),
+			'channels' => explode(',', (string) env('LOG_STACK', 'single')),
 			'ignore_exceptions' => false,
 		],
 
@@ -98,10 +98,10 @@ return [
 			'driver' => 'monolog',
 			'level' => env('LOG_LEVEL', 'debug'),
 			'handler' => StreamHandler::class,
-			'formatter' => env('LOG_STDERR_FORMATTER'),
-			'with' => [
+			'handler_with' => [
 				'stream' => 'php://stderr',
 			],
+			'formatter' => env('LOG_STDERR_FORMATTER'),
 			'processors' => [PsrLogMessageProcessor::class],
 		],
 
