@@ -24,7 +24,8 @@ return [
 	| used by your application. An example configuration is provided for
 	| each backend supported by Laravel. You're also free to add more.
 	|
-	| Drivers: "sync", "database", "beanstalkd", "sqs", "redis", "null"
+	| Drivers: "sync", "database", "beanstalkd", "sqs", "redis",
+	|          "deferred", "background", "failover", "null"
 	|
 	*/
 
@@ -70,6 +71,22 @@ return [
 			'retry_after' => (int) env('REDIS_QUEUE_RETRY_AFTER', 90),
 			'block_for' => null,
 			'after_commit' => false,
+		],
+
+		'deferred' => [
+			'driver' => 'deferred',
+		],
+
+		'background' => [
+			'driver' => 'background',
+		],
+
+		'failover' => [
+			'driver' => 'failover',
+			'connections' => [
+				'database',
+				'deferred',
+			],
 		],
 
 	],
