@@ -15,7 +15,11 @@ return [
 	*/
 
 	'postmark' => [
-		'token' => env('POSTMARK_TOKEN'),
+		'key' => env('POSTMARK_API_KEY'),
+	],
+
+	'resend' => [
+		'key' => env('RESEND_API_KEY'),
 	],
 
 	'ses' => [
