@@ -14,5 +14,6 @@ if (file_exists($maintenance = __DIR__ . '/../storage/framework/maintenance.php'
 require __DIR__ . '/../vendor/autoload.php';
 
 // Bootstrap Laravel and handle the request...
-(require_once __DIR__ . '/../bootstrap/app.php')
-	->handleRequest(Request::capture());
+$app = include_once __DIR__ . '/../bootstrap/app.php';
+
+$app->handleRequest(Request::capture());
