@@ -1,4 +1,4 @@
-import { mockServerError } from '../../support/commands';
+import { mockServerError } from '../../support/commands.js';
 
 describe('register', () => {
 	describe('with username that is taken', () => {
@@ -76,52 +76,53 @@ describe('register', () => {
 			// Verify email.
 			// TODO: With expired token on page load.
 			// TODO: With expired token on submit.
-			cy.visit(Cypress.env('mail_url'));
-			cy.contains(`[${Cypress.env('site_name')}] Verify Email Address`).click();
-			cy.get('#nav-plain-text-tab').click();
-			cy.get('[href*="/verify-email"]')
-				.then(($a) => {
-					// TODO: With invalid token.
-					// TODO: With invalid signature.
-					cy.visit($a.attr('href'));
-					cy.get('[data-cy="verify"]').click();
-					cy.closeToast('Email verified successfully.');
+			cy.origin(Cypress.env('mail_url'), () => {
+				cy.visit(Cypress.env('mail_url'));
+				cy.contains(`[${Cypress.env('site_name')}] Verify Email Address`).click();
+				cy.get('#nav-plain-text-tab').click();
+				cy.get('[href*="/verify-email"]').invoke('attr', 'href');
+			}).then((verifyEmailUrl) => {
+				// TODO: With invalid token.
+				// TODO: With invalid signature.
+				cy.visit(verifyEmailUrl);
+				cy.get('[data-cy="verify"]').click();
+				cy.closeToast('Email verified successfully.');
 
-					// Login.
-					cy.get('[name="username"]').type(username);
-					cy.get('[name="password"]').type(Cypress.env('default_password'));
-					cy.get('[type="submit"]').click();
-					cy.wait('@login').its('response.statusCode').should('equal', 200);
-					cy.location('pathname').should('eq', Cypress.env('public_path'));
+				// Login.
+				cy.get('[name="username"]').type(username);
+				cy.get('[name="password"]').type(Cypress.env('default_password'));
+				cy.get('[type="submit"]').click();
+				cy.wait('@login').its('response.statusCode').should('equal', 200);
+				cy.location('pathname').should('eq', Cypress.env('public_path'));
 
-					// Set measurement units.
-					cy.get('[data-cy="profile"]').click();
-					cy.get('#height + .formosa-suffix').should('contain', 'centimetres');
-					cy.get('[id="weight.weight"] + .formosa-suffix').should('contain', 'kgs');
-					cy.get('#measurement_units').select('imperial (weight in pounds, height in inches)');
-					cy.get('#height + .formosa-suffix').should('contain', 'inches');
-					cy.get('[id="weight.weight"] + .formosa-suffix').should('contain', 'lbs');
-					cy.get('#save-bmi').click();
-					cy.get('#measurement_units').should('be.disabled');
+				// Set measurement units.
+				cy.get('[data-cy="profile"]').click();
+				cy.get('#height + .formosa-suffix').should('contain', 'centimetres');
+				cy.get('[id="weight.weight"] + .formosa-suffix').should('contain', 'kgs');
+				cy.get('#measurement_units').select('imperial (weight in pounds, height in inches)');
+				cy.get('#height + .formosa-suffix').should('contain', 'inches');
+				cy.get('[id="weight.weight"] + .formosa-suffix').should('contain', 'lbs');
+				cy.get('#save-bmi').click();
+				cy.get('#measurement_units').should('be.disabled');
 
-					// Shows correct weight units on diary.
-					cy.get('.nav__link').contains('Diary').click();
-					cy.get('#weight + .formosa-suffix').should('contain', 'lbs');
+				// Shows correct weight units on diary.
+				cy.get('.nav__link').contains('Diary').click();
+				cy.get('#weight + .formosa-suffix').should('contain', 'lbs');
 
-					// Shows correct weight units on calendar.
-					cy.setWeight('130.5');
-					cy.get('.nav__link').contains('Calendar').click();
-					cy.wait('@getCalendar').its('response.statusCode').should('equal', 200);
-					cy.get('.calendar__item').contains('130.5 lbs').should('be.visible');
+				// Shows correct weight units on calendar.
+				cy.setWeight('130.5');
+				cy.get('.nav__link').contains('Calendar').click();
+				cy.wait('@getCalendar').its('response.statusCode').should('equal', 200);
+				cy.get('.calendar__item').contains('130.5 lbs').should('be.visible');
 
-					// Delete.
-					cy.get('[data-cy="profile"]').click();
-					cy.wait('@getUser').its('response.statusCode').should('equal', 200);
-					cy.get('.formosa-button--danger').contains('Delete account').click();
-					cy.get('dialog .formosa-button--danger').contains('Delete').click();
-					cy.wait('@deleteUser').its('response.statusCode').should('equal', 204);
-					cy.location('pathname').should('eq', Cypress.env('public_path'));
-				});
+				// Delete.
+				cy.get('[data-cy="profile"]').click();
+				cy.wait('@getUser').its('response.statusCode').should('equal', 200);
+				cy.get('.formosa-button--danger').contains('Delete account').click();
+				cy.get('dialog .formosa-button--danger').contains('Delete').click();
+				cy.wait('@deleteUser').its('response.statusCode').should('equal', 204);
+				cy.location('pathname').should('eq', Cypress.env('public_path'));
+			});
 		});
 	});
 
